Extend bootstrapper to initialize git repo of new project
Since these are new projects, they will not have a git repo. Initializes a
local repo for the project.

Required extensive modification to GitRepo. Previously GitRepo required
a remoteURL, but for an initial local project the remote URL is unneeded
(and likely unknown). GitRepo’s remoteURL is now an optional, which
requires additional checks within the class to support different scenarios.