Extend bootstrapper to add & commit all files in the directory
This required a refactor of GitRepo from a global Process to a local Process that gets
created when the GitAction is executed.  As per Apple documentation:

>An NSTask object can only be run once. Subsequent attempts to run the task raise an error.

a global process would require a new GitRepo instance for any operation performed on
the same repository.  This way multiple subsequent actions can be performed on the repo.