Refactor bootstrap for support of tech stack specific actions
Incorporates a new BootStrapConfig protocol that can be used to define a basic method
for a tech stack (now defined via a new enum) to do specific actions for that stack. Did
a major refactor moving a bunch of nested if statements to guard statements in an attempt to
improve flow within the main function.