download Android Jumpstart project to temporary folder.
The bootstrap project is creating it’s own git repo, so while we’re pulling code
from the jumpstart project, we don’t want to clone or check it out directly into
the bootstrap directory structure to avoid git history conflicts. Cloning to a temp
folder allows us to extract the information that we want and then copy it to the right
locations.