Makes ReadMe setup only happen for folders that are empty
We want to maintain the project file structure within git, but git only stores folders
if they are non-empty. As some folders for a project will only be populated as the
project progresses, we need to add some sort of file to maintain that structure initially.
Only need to add files to folders that are not populated by the tech stack bootstrapper,
so we will only add those ReadMe files after the bootstrapper has run.