Setup initial Android boostrapper that installs the jumpstart
Leverages the jumpstart project that Molded Bits has already written and is maintaining.
chaitools will clone the jumpstart project to a temp directory and then copy the correct
files to the right location.  This way we can maintain the jumpstarter without the .git
folders from the jumpstarter repo.

Molded Bits has some plans on how to address and improve the jumpstarter process
so will begin to maintain the AndroidBootstrap themselves.