Add verbosity levels to messaging.
chaitools in its current form is extremely chatty. This will replace all print statements
with statements within a new MessageTools structure that has four tiers of verbosity.
All statements default to a normal level of verbosity.  The BootstrapCommand now takes
in a value that allows you to specify what verbosity the system should print at.

Silent verbosity should be used only for those statements that *must* be printed for the
system to work. Verbose would be for a second tier of communication, Debug should be
used for debugging information that a normal user won’t understand (or for output from
third party processes that chaitools calls (like git).