Remove xcodeproj from source
For this project we are using Swift Package Manager and the Swift
build tools instead of xcode-build.  Any changes made to the Swift
package will require regenerating the xcodeproj file, so it doesn't
make sense to keep it in source control.  Added an exclude on the
project file to gitignore will allow individual devs to generate one
without worrying about it cluttering up the commits.