Add fastlane support
This is a very stripped down version of the Fastlane files that we
use for iOS projects.  It has one lane 'test' that will call
'swift test' and show the results. This should later be modified to
include a push to GitHub to publish a new version of the tap, because
that would be awesome.