Install javascript support libraries via yarn
We are standardizing on yarn as a package manager over NPM, so the initializer
Should follow the same standards. This will setup everything for ember, react, and
React-native. React itself doesn’t have a CLI like Ember and React Native, so I don’t
Know if installing it here actually helps anything, but it doesn’t hurt to be consistent.

Yarn installation requires updating the user’s shell profile. As we don’t have a standard
Shell to use, this could get complicated so I’ve punted on it for now and just have a message
Prompt to update the shell. Should be revisited.