Add support for gem installations
Right now this installs fastlane, bundler, and rails. Most other gems
will be installed with the Gemfile for the particular project. Unlike the
Other sets, this does not check for if the gem is installed before running
As the best way to get updates is by installing the gem again.