Fix linker issue for release builds
The release of 0.4.0 failed because the code wouldn't build in release.
This was due to a weird bug linking ChaiCommandKit to ChaiToolsKit that
only manifested using swift build -c release (vs a debug swift build).
Ultimately, the problem was the ChaiCommand protocol being private even
though there were public implementations of it that were defined in
ChaiCommandKit but used in ChaiToolsKit. This is a known issue
of the swift compiler discussed here: https://bugs.swift.org/browse/SR-2925.